requirements.txt
primer3-py==2.2.0
ispcr==0.9.1
pysam=0.23.3
pandas
sqlite